quick commit

+++ app/controllers/api/v1/subscriptions_controller.rb
@@ -9,10 +9,11 @@ module Api
       
       def show
         if params[:clientKey] == "OIUDFBOSU097098Y34IUBSFDV09898dvhsodfkjbaf93tb93rbrv"
-            subscription= Subscription.find_all_by(user: params[:user])
+            subscription= Subscription.where(user: params[:user])
+            #subscription= Subscription.find_by(user: params[:user])
             render json: {status:'SUCCESS',message:'Loaded subscription',data:subscription},status: :ok
         else
-            render json: {status:'FAILURE',message:'Incorrect Password'},status: :bad_password
+            render json: {status:'FAILURE ON SHOW',message:'Incorrect Password'},status: :bad_password
         end
 
       end
@@ -20,17 +21,29 @@ module Api
       def create
         subscription = Subscription.new(subscription_params)
 
-        if subscription.save
-          render json: {status: 'SUCCESS', message: 'Saved Subscription', data:subscription},status: :ok
+        if params[:clientKey] == "OIUDFBOSU097098Y34IUBSFDV09898dvhsodfkjbaf93tb93rbrv"
+          if subscription.save
+            render json: {status: 'SUCCESS', message: 'Saved Subscription', data:subscription},status: :ok
+          else
+            render json: {status: 'ERROR', message: 'Subscription not saved', data:subscription.error},status: :unprocessable_entity
+          end
         else
-          render json: {status: 'ERROR', message: 'Subscription not saved', data:subscription.error},status: :unprocessable_entity
+            render json: {status:'FAILURE ON SHOW',message:'Incorrect Password'},status: :bad_password
         end
+
       end
       
       def destroy
-        subscription = Subscription.find(params[:id])
-        subscription.destroy
-        render json: {status: 'SUCCESS', message: 'Subscription deleted successfully', data:subscription},status: :ok
+        if params[:clientKey] == "OIUDFBOSU097098Y34IUBSFDV09898dvhsodfkjbaf93tb93rbrv"
+          #subscription = Subscription.where(user: params[:user], topic: params[:issue])
+          subscription = Subscription.find_by(user: params[:user], issue: params[:issue])
+          subscription.destroy
+          render json: {status: 'SUCCESS', message: 'Subscription deleted successfully', data:subscription},status: :ok
+      
+        else
+            render json: {status:'FAILURE ON SHOW',message:'Incorrect Password'},status: :bad_password
+        end
+
       end
 
       private
